+++ Cognite.DataProcessing/Cognite.DataProcessing.csproj
@@ -1,0 +1,26 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+	<GenerateDocumentationFile>true</GenerateDocumentationFile>
+	<PackageId>Cognite.DataProcessing</PackageId>
+	<Author>Cognite AS</Author>
+	<Company>Cognite AS</Company>
+	<Copyright>Cognite AS</Copyright>
+	<PackageLicenseFile>LICENSE</PackageLicenseFile>
+	<RootNamespace>Cognite.DataProcessing</RootNamespace>
+	<Description>
+		A library containing utility functions for data processing
+	</Description>
+	<EnableNETAnalyzers>true</EnableNETAnalyzers>
+	<AnalysisMode>Default</AnalysisMode>
+	<EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
+  </PropertyGroup>
+  <ItemGroup>
+	<None Include="..\LICENSE" Pack="true" Visible="false" PackagePath="" />
+  </ItemGroup>
+  <ItemGroup>
+	<PackageReference Include="MathNet.Numerics" Version="5.0.0" />
+  </ItemGroup>
+
+</Project>
