+++ Cognite.Simulator.Tests/DataProcessingTests/DetectorsTest.cs
@@ -1,7 +1,7 @@
 using Xunit;
-using Cognite.Simulator.Utils;
+using Cognite.DataProcessing;
 
-namespace Cognite.Simulator.Tests.UtilsTests;
+namespace Cognite.DataProcessing.Tests;
 
 public class DetectorsTest
 {
