+++ Cognite.Simulator.sln
@@ -13,7 +13,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.runsettings = .runsettings
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cognite.Simulator.Tests", "Cognite.Simulator.Tests\Cognite.Simulator.Tests.csproj", "{A6584AFE-EB3B-4AEB-8532-B46594B51323}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cognite.Simulator.Tests", "Cognite.Simulator.Tests\Cognite.Simulator.Tests.csproj", "{A6584AFE-EB3B-4AEB-8532-B46594B51323}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cognite.DataProcessing", "Cognite.DataProcessing\Cognite.DataProcessing.csproj", "{64F2199F-B74C-4AF7-8E5D-3999302F887D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,6 +35,10 @@ Global
 		{A6584AFE-EB3B-4AEB-8532-B46594B51323}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A6584AFE-EB3B-4AEB-8532-B46594B51323}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A6584AFE-EB3B-4AEB-8532-B46594B51323}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64F2199F-B74C-4AF7-8E5D-3999302F887D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64F2199F-B74C-4AF7-8E5D-3999302F887D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64F2199F-B74C-4AF7-8E5D-3999302F887D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64F2199F-B74C-4AF7-8E5D-3999302F887D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
